Rename lifecycle to events where it makes sense
