+++ examples/README.md
@@ -3,4 +3,3 @@
 1. [Backwards compat](./backwards-compat) shows an example of **Netlify Build** using a `netlify.toml` file with no
    changes
 2. [Example one](./example-one) shows how to use plugins & new `netlify.yml`
-3. [Example two](./example-two) expands on example on and uses multiple plugins & build lifecycle commands
