+++ packages/config/src/events.js
