+++ scripts/docs.js
@@ -8,7 +8,7 @@ const request = require('sync-request')
 
 const ROOT_DIR = path.join(__dirname, '..')
 const CONSTANTS_PATH = path.join(ROOT_DIR, 'packages/build/src/plugins/child', 'constants.js')
-const LIFECYCLE_PATH = path.join(ROOT_DIR, 'packages/config/src/lifecycle.js')
+const LIFECYCLE_PATH = path.join(ROOT_DIR, 'packages/config/src/events.js')
 const PLUGINS_DATABASE_URL = 'https://raw.githubusercontent.com/netlify/plugins/master/plugins.json'
 const PLUGIN_NAME_REGEX = /(?:(?:^|-)netlify-plugin(?:-|$))|(?:(?:^|-)netlify(?:-|$))/
